kategoriler eklendi

+++ BlogFall/Areas/Admin/Controllers/PostsController.cs
@@ -1,14 +1,18 @@
 ﻿using BlogFall.Areas.Admin.ViewModels;
+using BlogFall.Attributes;
 using BlogFall.Models;
 using Microsoft.AspNet.Identity;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Net;
 using System.Web;
 using System.Web.Mvc;
 
 namespace BlogFall.Areas.Admin.Controllers
 {
+    [Breadcrumb("Yazılar")]
     public class PostsController : AdminBaseController
     {
         // GET: Admin/Posts
@@ -21,6 +25,10 @@ namespace BlogFall.Areas.Admin.Controllers
         public ActionResult Delete(int id)
         {
             var post = db.Posts.Find(id);
+            if (post == null)
+            {
+                return HttpNotFound();
+            }
             db.Posts.Remove(post);
             db.SaveChanges();
 
@@ -43,6 +51,7 @@ namespace BlogFall.Areas.Admin.Controllers
         [HttpPost]
         [ValidateInput(false)]
         [ValidateAntiForgeryToken]
+        [Breadcrumb("Düzenle")]
         public ActionResult Edit(PostEditViewModel model)
         {
             if (ModelState.IsValid)
@@ -88,6 +97,24 @@ namespace BlogFall.Areas.Admin.Controllers
             }
             ViewBag.CategoryId = new SelectList(db.Categories.ToList(), "Id", "CategoryName");
             return View("Edit", new PostEditViewModel());
+        }
+
+        [HttpPost]
+        public ActionResult AjaxImageUpload(HttpPostedFileBase file)
+        {
+            if (file == null || file.ContentLength == 0 || !file.ContentType.StartsWith("image/"))
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+
+            var saveFolderPath = Server.MapPath("~/Upload/Posts");
+            var ext = Path.GetExtension(file.FileName);
+            var saveFileName = Guid.NewGuid() + ext;
+            var saveFilePath = Path.Combine(saveFolderPath, saveFileName);
+
+            file.SaveAs(saveFilePath);
+            return Json(new { url = Url.Content("~/Upload/Posts/" + saveFileName) });
+        
         }
     }
 }
